Might need to push this to the other scripts...
Ran into issue on rebuild where the first `chown` of /var/www/redmine was
failing with an `chown: fts_read failed: No such file or directory` even
though the directory and all files exists.

Assuming that issue only manifests on re-builds due to the some of the
sub-dirs being EFS-hosted and autofs-managed leading to odd time-outs.
Changing to the find-based method appears to work around whatever the
true underlying problem is.